+++ src/studio/lib/data_sources/data_sources.dart
@@ -1,5 +1,3 @@
-library data_sources;
-
 export 'base.dart';
 export 'file_source.dart';
 export 'bundle_source.dart';