+++ src/studio/packages/qtadmin-org/lib/org_barrel.dart
@@ -1,5 +1,0 @@
-library qtadmin_org;
-
-export 'org.dart';
-export 'src/screens/org_screen.dart';
-export 'src/views/stat_item.dart';